requirements.txt
boto3
flask
flask-cors
flask-restx
okta
okta-jwt-verifier
python-dotenv
pydantic
sendgrid
twilio
pymongo[srv]
Flask-PyMongo